backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0
python-engineio>=4.8.0
marshmallow>=3.20.0
python-dotenv>=1.0.0
modal>=0.68.0
httpx>=0.28.0
gunicorn>=21.2.0
eventlet>=0.37.0
